refactor: create authorization dir
